Move aiosnow.model to aiosnow.models.common
