+++ Makefile
@@ -45,5 +45,5 @@ lint:
 reformat:
 	poetry run autoflake --in-place --recursive aiosnow tests
 	poetry run black aiosnow tests examples
-	poetry run isort --multi-line=3 --trailing-comma --force-grid-wrap=0 --combine-as --line-width 88 --recursive --apply aiosnow tests
+	poetry run isort --multi-line=3 --trailing-comma --force-grid-wrap=0 --combine-as --line-width 88 aiosnow tests
