+++ aiosnow/__init__.py
@@ -1,4 +1,3 @@
 from .client import Client
-from .model.base import BaseModel
-from .model.schema import PartialSchema, Pluck, fields
+from .models.common import BaseModel, PartialSchema, Pluck, fields
 from .query import QueryBuilder, select
