+++ aiosnow/models/common/base.py
