+++ aiosnow/models/common/schema/base.py
