+++ aiosnow/models/common/schema/partial.py
