+++ tests/client/test_app_session_input.py
@@ -7,7 +7,7 @@ from aiosnow.request.response import Response
 from aiosnow.session import Session
 
 
-def test_app_session_invalid_type():
+def test_client_session_invalid_type():
     """Invalid object type passed to `session` should raise an InvalidSessionType exception"""
 
     fail_str = dict(address="test.service-now.com", session="test")
@@ -18,7 +18,7 @@ def test_app_session_invalid_type():
         Client(**fail_int)
 
 
-def test_app_session_mutual_exclusive():
+def test_client_session_mutual_exclusive():
     """Passing both a session and session factory configuration should raise ConfigurationException"""
 
     session = Session()
@@ -30,14 +30,14 @@ def test_app_session_mutual_exclusive():
         Client("test.service-now.com", session=session, verify_ssl=True)
 
 
-def test_app_session_no_auth_method():
+def test_client_session_no_auth_method():
     """No authentication method to Client should raise ConfigurationException"""
 
     with pytest.raises(ConfigurationException):
         Client("test.service-now.com")
 
 
-def test_app_session_config_full():
+def test_client_session_config_full():
     """Session config should load into a config object"""
 
     config = dict(
@@ -47,23 +47,23 @@ def test_app_session_config_full():
         verify_ssl=True,
     )
 
-    app = Client(**config)
+    snow = Client(**config)
 
-    assert app.config.address == config["address"]
-    assert app.config.session.basic_auth == config["basic_auth"]
-    assert app.config.session.use_ssl == config["use_ssl"]
-    assert app.config.session.verify_ssl == config["verify_ssl"]
+    assert snow.config.address == config["address"]
+    assert snow.config.session.basic_auth == config["basic_auth"]
+    assert snow.config.session.use_ssl == config["use_ssl"]
+    assert snow.config.session.verify_ssl == config["verify_ssl"]
 
 
-def test_app_session_object():
+def test_client_session_object():
     """Compatible user-provided Session objects should be returned from Client.get_session"""
 
     session = Session()
-    app = Client("test.service-now.com", session=session)
-    assert app.get_session() == app._preconf_session == session
+    snow = Client("test.service-now.com", session=session)
+    assert snow.get_session() == snow._preconf_session == session
 
 
-def test_app_session_invalid_response_class():
+def test_client_session_invalid_response_class():
     """Compatible user-provided Session objects should be returned from Client.get_session"""
 
     session = aiohttp.ClientSession()
@@ -72,10 +72,10 @@ def test_app_session_invalid_response_class():
         Client("test.service-now.com", session=session)
 
 
-def test_app_session_response_class():
+def test_client_session_response_class():
     """Compatible user-provided Session objects should be returned from Client.get_session"""
 
     session = aiohttp.ClientSession(response_class=Response)
-    app = Client("test.service-now.com", session=session)
+    snow = Client("test.service-now.com", session=session)
 
-    assert app.get_session()._response_class == Response
+    assert snow.get_session()._response_class == Response
